refactor(@leav/ui): extract deactivateAction to be ready for new default item actions
