+++ libs/ui/src/components/Explorer/Explorer.tsx
@@ -1,11 +1,8 @@
 // Copyright LEAV Solutions 2017
 // This file is released under LGPL V3
 // License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-import {ComponentProps, FunctionComponent, useState} from 'react';
-import {FaTrash} from 'react-icons/fa';
-import {KitModal} from 'aristid-ds';
-import {useSharedTranslation} from '_ui/hooks/useSharedTranslation';
-import {useDeactivateRecordsMutation} from '_ui/_gqlTypes';
+import {FunctionComponent} from 'react';
+import {useDeactivateAction} from './useDeactivateAction';
 import {DataView} from './DataView';
 import {useExplorerData} from './useExplorerData';
 import {ItemActions} from './types';
@@ -13,64 +10,44 @@ import {ItemActions} from './types';
 interface IExplorerProps {
     library: string;
     itemActions: ItemActions<any>;
-    defaultActionsForItem?: [] | ['deactivate'] | ['edit'] | ['deactivate', 'edit'] | undefined;
+    defaultActionsForItem?:
+        | []
+        | ['deactivate']
+        | ['edit']
+        | ['edit', 'deactivate']
+        | ['deactivate', 'edit']
+        | undefined;
 }
+
+const isNotEmpty = <T extends unknown[]>(union: T): union is Exclude<T, []> => union.length > 0;
 
 export const Explorer: FunctionComponent<IExplorerProps> = ({
     library,
     itemActions,
-    defaultActionsForItem = ['deactivate', 'edit']
+    defaultActionsForItem = ['edit', 'deactivate']
 }) => {
-    const {t} = useSharedTranslation();
-
     const {data, loading} = useExplorerData(library); // TODO: refresh when go back on page
-    const [deactivatedItemIds, setDeactivatedItemIds] = useState<string[]>([]);
 
-    const [deactivateRecords] = useDeactivateRecordsMutation();
-    const _deactivateItem = async ({itemId, libraryId}) => {
-        await deactivateRecords({
-            variables: {
-                libraryId,
-                recordsIds: [itemId]
-            }
-        });
+    const isDeactivateActionEnabled =
+        isNotEmpty(defaultActionsForItem) &&
+        defaultActionsForItem.some(
+            // TODO: try with Array.includes better typing
+            a => a === 'deactivate'
+        );
 
-        setDeactivatedItemIds([...deactivatedItemIds, itemId]);
-    };
-
-    const _deleteAction: ItemActions<any>[number] = {
-        label: t('explorer.deactivateItem'),
-        icon: <FaTrash />,
-        isDanger: true,
-        callback: item =>
-            KitModal.confirm({
-                type: 'confirm',
-                dangerConfirm: true,
-                content: t('records_deactivation.confirm_one') ?? undefined,
-                okText: t('global.submit') ?? undefined,
-                cancelText: t('global.cancel') ?? undefined,
-                onOk: () => _deactivateItem(item)
-            })
-    };
-
-    const actionsOnItem: ComponentProps<typeof DataView>['itemActions'] = defaultActionsForItem?.some(
-        // TODO: try with Array.includes better typing
-        a => a === 'deactivate'
-    )
-        ? [_deleteAction, ...itemActions]
-        : itemActions;
+    const {deactivateAction, deactivatedItemIds} = useDeactivateAction(isDeactivateActionEnabled);
 
     return (
-        <div>
+        <>
             {loading ? (
                 'Loading...'
             ) : (
                 <DataView
                     dataGroupedFilteredSorted={data?.filter(({itemId}) => !deactivatedItemIds.includes(itemId)) ?? []}
                     attributesToDisplay={['itemId', 'whoAmI']}
-                    itemActions={actionsOnItem}
+                    itemActions={[deactivateAction, ...itemActions].filter(action => !!action)}
                 />
             )}
-        </div>
+        </>
     );
 };
